QB-Cocaine - Bug fixes and all peds and vehicle creation is handled client-side rather than server
Major changes
So now that the peds and vehicle is created client-side and they are not networked, meaning synced on the network.
Other players cannot see buyer animate for another player he's selling cocaine.
It is a price i am willing to pay as the script is more stable and works better.

Other changes
Positions of peds and vehicle changed in `config.lua`